requirements.txt
tinify>=1.6.0
Pillow>=9.0.0